fix: keep decorator metadata runtime portable
Verified: precommit-verify ✕
