+++ packages/decorator-metadata/src/runtime/decorators.lib.ts
@@ -1,7 +1,7 @@
 import { isClassConstructor, toUnknownCallable } from '@zeltjs/unsafe-type-lib';
 import { match, P } from 'ts-pattern';
 
-import { captureStackTrace, withWrapperFiles } from './position.lib';
+import { captureStackTrace, withCallStackTrace } from './trace.lib';
 import {
   aggregateMembers,
   ensureClassMeta,
@@ -212,7 +212,7 @@ export const createClassDecorator = <TProps extends object, E extends Error = Er
       const err: E | undefined = rejectIfApplied(getClassMetadata(cls)?.props ?? []);
       if (err) throw err;
     }
-    const trace = withWrapperFiles(defineTrace, captureStackTrace());
+    const trace = withCallStackTrace(defineTrace, captureStackTrace());
     recordClass(cls, trace, props);
     aggregateMembers(cls, classKey);
     afterApply?.(cls);
@@ -232,7 +232,7 @@ export const createMethodDecorator = <TProps extends object, E extends Error = E
       throw err;
     }
 
-    const trace = withWrapperFiles(defineTrace, captureStackTrace());
+    const trace = withCallStackTrace(defineTrace, captureStackTrace());
     recordMethod(info.classKey, info.name, trace, props);
     if (options?.afterApply && info.method) options.afterApply(info.method, info.name);
   });
@@ -249,7 +249,7 @@ export const createPropertyDecorator = <TProps extends object>(
   const defineTrace = captureStackTrace();
 
   return adaptPropertyContext((classKey, name) => {
-    const trace = withWrapperFiles(defineTrace, captureStackTrace());
+    const trace = withCallStackTrace(defineTrace, captureStackTrace());
     recordProperty(classKey, name, trace, props);
     options?.afterApply?.(name);
   });
@@ -271,7 +271,7 @@ export const composeClassDecorators = (...decorators: ClassDecoratorFn[]): Class
     const cls = asObject(args[0]);
     if (!cls) return undefined;
 
-    const trace = withWrapperFiles(defineTrace, captureStackTrace());
+    const trace = withCallStackTrace(defineTrace, captureStackTrace());
     if (trace) ensureClassMeta(cls, trace);
 
     for (const dec of decorators) {
