+++ packages/decorator-metadata/src/runtime/index.ts
@@ -14,8 +14,8 @@ export {
   createMethodDecorator,
   createPropertyDecorator,
 } from './decorators.lib';
-export type { Position, ResolvePositionOptions, StackTrace } from './position.lib';
-export { captureStackTrace, resolvePosition, withWrapperFiles } from './position.lib';
+export type { StackTrace } from './trace.lib';
+export { captureStackTrace, withCallStackTrace } from './trace.lib';
 export type { ClassMeta, MethodMeta, PropertyMeta } from './store.lib';
 export {
   aggregateMembers,
